+++ globalutils.pas
@@ -7,14 +7,14 @@ unit globalutils;
 interface
 
 (* Select random number from a range *)
-function randomRange(fromNumber, toNumber: integer): integer;
+function randomRange(fromNumber, toNumber: smallint): smallint;
 
 implementation
 
 // Random(Range End - Range Start) + Range Start;
-function randomRange(fromNumber, toNumber: integer): integer;
+function randomRange(fromNumber, toNumber: smallint): smallint;
 var
-  p: integer;
+  p: smallint;
 begin
   p := toNumber - fromNumber;
   Result := random(p + 1) + fromNumber;
